https://open.spotify.com/track/0ttrMR1b0PXdDDPeYPBEOf?si=e68aa4fc5b5e421f

+++ functions/controllers/TeacherController.js
@@ -3,12 +3,12 @@ const HelperRepository = require("../repositories/HelperRepository");
 const StudentRepository = require("../repositories/StudentRepository");
 const TeacherRepository = require("../repositories/TeacherRepository");
 class TeacherController {
-    static async createClass(req, res){
+    static async createClassForTheTeacher(req, res){
         try {
             const { className } = req.body;
             console.log("Teacher ID: ", req.uid)
             const classId = await ClassRepository.createClass(className, req.uid)
-            TeacherRepository.addClass(classId, req.uid)
+            await TeacherRepository.addClass(classId, req.uid)
             res.status(200).json({ message: "Success!"});
         } catch (error) {
             console.error("Error creating class:", error);
@@ -18,11 +18,16 @@ class TeacherController {
 
     static async addStudentToClass(req, res) {
         try {
-            const { classId, emailOrName } = req.body;
-            const { studentUid, classData, classRef } = await HelperRepository.studentUidAndClassData(classId, emailOrName);
+            const { className, studentName } = req.body;
+            const { studentUid, classData, classRef, classId } = await HelperRepository.studentUidAndClassData(className, studentName);
     
             const studentUids = classData.studentUids || [];
 
+            const studentClassChecker = await StudentRepository.studentClassExistChecker(studentUid);
+            if(studentClassChecker == true){
+                throw new Error(`Student already have a Class!`);
+            } 
+            
             if (!studentUids.includes(studentUid)) {
                 await StudentRepository.addClass(studentUid, classId)
 
@@ -39,9 +44,15 @@ class TeacherController {
 
     static async removeStudentFromClass(req, res) {
         try {
-            const { classId, emailOrName } = req.body;
-            const { studentUid, classData, classRef } = await HelperRepository.studentUidAndClassData(classId, emailOrName);
+            const { className, studentName } = req.body;
+            const { studentUid, classData, classRef, classId } = await HelperRepository.studentUidAndClassData(className, studentName);
     
+            const studentClassChecker = await StudentRepository.studentClassExistChecker(studentUid);
+
+            if(studentClassChecker == false){
+                throw new Error(`Student does not have Class!`);
+            } 
+
             const updatedUids = (classData.studentUids || []).filter(uid => uid !== studentUid);
             
             await StudentRepository.removeClass(studentUid);
@@ -51,6 +62,30 @@ class TeacherController {
         } catch (error) {
             console.error("Error removing student:", error);
             res.status(500).json({ error: "Failed to remove student." });
+        }
+    }
+
+    static async removeClassOfTheTeacher(req, res){
+        try {
+            const { className } = req.body;
+            if (!className) throw new Error("Class name is required.");
+    
+            const classId = await TeacherRepository.removeClass(req.uid, className);
+            const classData = await ClassRepository.getClassData(classId);
+            const studentUids = classData.studentUids || [];
+    
+            await Promise.all(
+                studentUids.map(async (studentUid) =>
+                    await StudentRepository.removeClass(studentUid)
+                )
+            );
+    
+            await ClassRepository.removeClassFromDocuments(classId);
+    
+            res.status(200).json({ message: "Success!" });
+        } catch (error) {
+            console.error("Error removing class:", error);
+            res.status(500).json({ error: "Failed to remove class." });
         }
     }
